cleaned up useless shit

+++ main.c
@@ -6,7 +6,7 @@
 /*   By: dromansk <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/13 14:34:29 by dromansk          #+#    #+#             */
-/*   Updated: 2018/12/13 14:36:49 by dromansk         ###   ########.fr       */
+/*   Updated: 2019/01/02 17:28:10 by dromansk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 #include <stdio.h>
 
 int		main(void){
-	double pi;
+	long double pi;
 
 	pi = 3.141592653589793;
-	printf("expected %Lf\n", pi);
-	ft_printf("result   %Lf\n", pi);
+	printf("expected %f\n", (double)pi);
+	ft_printf("result   %f\n", pi);
 	return (0);}
